Split Android data local tests by package
